cdiscount_import
This Programm allows the user to access product data via the plentymarkets Api
and restructure that data to fit cdiscounts uploadfile requirements.
While doing so the Programm checks for any errors in the data that cdiscount wouldn't
take and puts those into an extra list, so the user gets an overview of the data he has to fix.

Signed-off-by: Jan <jan.sallermann@panasiam.de>