Refactor: Make parent sku comparision more meaningful
Stop using raw indeces for within a string comparision of two data
structures, in order to make the code more meaningful and easier to
understand. Transform the `texts` and `error_texts` lists to
dictionaries to access elements with a name instead of an index.
Implement a constant for the index of the parent sku field within
the variations list.

Signed-off-by: Sebastian Fricke <sebastian.fricke@posteo.net>