Fix: Fix parent SKU handling
Pull the actual SKU from the parent and not a parent SKU value stored
under variation SKUs from a fixed index. Assign the texts via the items
ID and not via a parent SKU, that resembles the item ID by chance in
case they are equal.
Rework the item_ids data-structure to be a dictionary instead of a list
and map every variation ID encountered during the data fetch, to the
appropriate item ID.

Signed-off-by: Sebastian Fricke <sebastian.fricke@posteo.net>