Refactor: Drop fixed access to category branch ID
The previous implementation fetched the branch ID of the default
category from a fixed index in the list. For the cdiscount market we
need the default category from a specific plenty system in case there
are multiple present. Get the plenty ID from the configuration and pull
the default category from the specific default category for that
plentymarkets instance.

Signed-off-by: Sebastian Fricke <sebastian.fricke@posteo.net>